# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: API Plataforma Web MOOC Académica
  description: Backend monolito modular en Go para la plataforma de cursos masivos abiertos en línea.
  version: 1.0.0
servers:
  - url: http://localhost:8081
    description: Entorno Docker Local
paths:
...
  /api/v1/admin/badges/{badgeId}/revoke:
    post:
      summary: Revocar una insignia emitida (Solo Admin)
      security:
        - BearerAuth: []
      parameters:
        - name: badgeId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Insignia revocada exitosamente
...
  /api/v1/media/presigned-upload:
    post:
      summary: Generar URL prefirmada para carga S3
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [object_key, content_type]
              properties:
                object_key:
                  type: string
                content_type:
                  type: string
      responses:
        '200':
          description: URL PUT prefirmada S3 retornada
  /api/v1/media/resources/{resourceId}/complete-upload:
    post:
      summary: Notificar finalización de carga y encolar transcodificación
      security:
        - BearerAuth: []
      parameters:
        - name: resourceId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [object_key]
              properties:
                object_key:
                  type: string
      responses:
        '200':
          description: Carga validada y worker asíncrono encolado
  /api/v1/learning/enrollments:
    post:
      summary: Inscribir estudiante a un curso
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [course_id]
              properties:
                course_id:
                  type: string
                  format: uuid
      responses:
        '200':
          description: Inscripción realizada exitosamente
  /api/v1/learning/heartbeat:
    post:
      summary: Enviar pulso de reproducción y permanencia
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [course_id, resource_stable_id, dwell_time_seconds]
              properties:
                course_id:
                  type: string
                  format: uuid
                resource_stable_id:
                  type: string
                  format: uuid
                dwell_time_seconds:
                  type: integer
                last_position_seconds:
                  type: integer
      responses:
        '200':
          description: Progreso actualizado en servidor
  /api/v1/badges/verify/{code}:
    get:
      summary: Verificar autenticidad pública de una insignia
      parameters:
        - name: code
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Detalles e integridad de la insignia
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Bearer